Added compiler option SINGLE to test single precision kernels and drivers. Make/CMake files are not updated yet.

+++ src/kernels/cuda/tcf/cuda_tcf_cp.h
@@ -12,8 +12,8 @@ void K_CUDA_TCF_CP_Lagrange(
     int call_type, int num_source, int num_cluster,
     int batch_num_sources, int batch_idx_start,
     int cluster_q_start, int cluster_pts_start, int interp_order_lim,
-    double *source_x, double *source_y, double *source_z, double *source_q,
-    double *cluster_x, double *cluster_y, double *cluster_z, double *cluster_q,
+    FLOAT *source_x, FLOAT *source_y, FLOAT *source_z, FLOAT *source_q,
+    FLOAT *cluster_x, FLOAT *cluster_y, FLOAT *cluster_z, double *cluster_q,
     struct RunParams *run_params);
 
 #ifdef __CUDACC__
